+++ src/keri/metric/__init__.py
@@ -1,0 +1,7 @@
+# -*- encoding: utf-8 -*-
+"""
+KERI
+keri.metric Package
+"""
+
+from .metricing import EscrowEnd
